Handle all wgrib2 calls to exit if they fail

+++ API/NBM_Fire_Local_Ingest.py
@@ -264,6 +264,9 @@ cmd = (
 )
 # Run wgrib2
 spOUT = subprocess.run(cmd, shell=True, capture_output=True, encoding="utf-8")
+if spOUT.returncode != 0:
+    print(spOUT.stderr)
+    sys.exit()
 
 # Check output from wgrib2
 # print(spOUT.stdout)
@@ -280,6 +283,9 @@ cmd2 = (
     + "_wgrib2_merged_order.grib"
 )
 spOUT2 = subprocess.run(cmd2, shell=True, capture_output=True, encoding="utf-8")
+if spOUT2.returncode != 0:
+    print(spOUT2.stderr)
+    sys.exit()
 os.remove(forecast_process_path + "_wgrib2_merged.grib2")
 
 # Convert to NetCDF
@@ -295,6 +301,10 @@ cmd4 = (
 
 # Run wgrib2 to rotate winds and save as NetCDF
 spOUT4 = subprocess.run(cmd4, shell=True, capture_output=True, encoding="utf-8")
+if spOUT4.returncode != 0:
+    print(spOUT4.stderr)
+    sys.exit()
+
 os.remove(forecast_process_path + "_wgrib2_merged_order.grib")
 
 #######
@@ -452,6 +462,9 @@ for i in range(hisPeriod, 1, -6):
         + "_wgrib2_merged_order.grib"
     )
     spOUT1 = subprocess.run(cmd1, shell=True, capture_output=True, encoding="utf-8")
+    if spOUT1.returncode != 0:
+        print(spOUT1.stderr)
+        sys.exit()
 
     # Convert to NetCDF
     cmd3 = (
@@ -464,6 +477,9 @@ for i in range(hisPeriod, 1, -6):
         + "_wgrib_merge.nc"
     )
     spOUT3 = subprocess.run(cmd3, shell=True, capture_output=True, encoding="utf-8")
+    if spOUT3.returncode != 0:
+        print(spOUT3.stderr)
+        sys.exit()
 
     # Merge the  xarrays
     # Read the netcdf file using xarray
